Add helper module and decrypt methods
