+++ test/encrypt_test.rb
@@ -56,16 +56,16 @@ class EncryptTest < Minitest::Test
   end
 
 
-  def test_shift_to_hash_method
-    expected = expected = {1 => "a", 2 => "b", 3 => "c", 4 => "d", 5 => "e", 6 => "f", 7 => "g", 8 => "h", 9 => "i", 10 => "j", 11 => "k", 12 => "l", 13 => "m", 14 => "n", 15 => "o", 16 => "p", 17 => "q", 18 => "r", 19 => "s" , 20  => "t", 21 => "u", 22 => "v", 23 => "w", 24 => "x", 25 => "y", 26 => "z", 27 => " ",
-    }
-    actual = @enigma.shifted
-    refute_equal  expected, actual
+  def test_encode_method
+    expected = "hello world"
+    actual = @enigma.encode
+    refute_equal expected, actual
   end
 
   def test_it_can_encrypt
+
     expected = {
-    encryption: @enigma.shifted,
+    encryption: @enigma.encode,
     key: "54321",
     date: "040895"
     }
